+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Used rowmapper for in memory db
+- Adds email verification APIs
+
 ## [3.0.1] - 2020-10-27
 ### Changed
 - Makes Hello API do a db query as well for better status checking
@@ -18,7 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.5.2] - 2020-10-25
 ### Fixed
 - Issue #84 - Correct access token signing key expiry not being sent by APIs 
->>>>>>> master
 
 ## [2.5.1] - 2020-10-08
 ### Changed